engine: Implement TestHostVolumeMount for Windows
